Crunchy should not irclink his own links from hubot
Prior to this commit, when somebody asked crunchy for something via the 'img
me' interface command set, crunchy would do the needful, but then also post the
returned URI to tumble as user crunchy. These links are usually more for fun
and less of an important part to tumble. So, this small fix disables that.